Tests and resulting refactor
